+++ Frontend/Assets/3DGamekit/Scripts/Game/Player/Inventory.cs
@@ -6,6 +6,6 @@ namespace Gamekit3D
 {
     public class Inventory
     {
-        public static Dictionary<String, Int16> player_Inventory = new Dictionary<string, Int16>();
+        public static Dictionary<String, int> player_Inventory = new Dictionary<string, int>();
     }
 }
